refactor: js->ts로 변환 완료
