+++ week3-ts/src/types/index.ts
